requirements.txt
python-statemachine

